worked on a bit of a refactor
