+++ server/router/admin.js
@@ -1,5 +1,5 @@
-const crud_move = require('../crud/move');
-const crud_user = require('../crud/user/user');
+const crud_user_basic = require('../crud/user/basic');
+const crud_admin = require('../crud/admin');
 const crud_interact = require('../crud/interact/interact');
 const config = require('../config');
 
@@ -13,14 +13,14 @@ function add_routes(socket, io) {
         var email_target = data["msg"].replace("teleport to", "").trim();
 
         // check if current user is admin
-        crud_user.is_admin(socket.id).catch(console.dir).then( (user) => {
+        crud_admin.is_admin(socket.id).catch(console.dir).then( (user) => {
             if(!user["admin"]) {
                 socket.send({data: "Only admin users can teleport"});
                 return;
             }
 
             // check if target user exists
-            crud_user.get_other_user(email_target).catch(console.dir).then( (user) => {
+            crud_user_basic.get_user_by_email(email_target).catch(console.dir).then( (user) => {
                 if(user === null) {
                     socket.send({data: "User not found: " + email_target});
                     return;
@@ -32,7 +32,7 @@ function add_routes(socket, io) {
                 var target_angle = (user["angle"] + Math.PI) // turn to face them
 
                 // move current user to target user
-                crud_move.teleport(
+                crud_admin.teleport(
                     socket, target_lat, target_long, target_angle
                 ).catch(console.dir).then( () => {
                     socket.send({data: "Teleport successful"});
@@ -50,20 +50,20 @@ function add_routes(socket, io) {
         var email_target = data["msg"].replace("ban", "").trim();
 
         // check if current user is admin
-        crud_user.is_admin(socket.id).catch(console.dir).then( (user) => {
+        crud_admin.is_admin(socket.id).catch(console.dir).then( (user) => {
             if(!user["admin"]) {
                 socket.send({data: "Only admin users can ban people"});
                 return;
             }
 
             // check if target user exists
-            crud_user.get_other_user(email_target).catch(console.dir).then( (user) => {
+            crud_user_basic.get_user_by_email(email_target).catch(console.dir).then( (user) => {
                 if(user === null) {
                     socket.send({data: "User not found: " + email_target});
                     return;
                 }
 
-                crud_user.ban(email_target, io).catch(console.dir).then( () => {
+                crud_admin.ban(email_target, io).catch(console.dir).then( () => {
                     socket.send({data: "User banned"});
                 });
             });
